# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Response:
    properties:
      data: {}
      status:
        $ref: '#/definitions/model.status'
    type: object
  model.status:
    enum:
    - success
    - error
    type: string
    x-enum-varnames:
    - success
    - err
...
  user.Login:
    properties:
      password:
        example: Qwerty123
        maxLength: 72
        minLength: 8
        type: string
      username:
        example: JackVorobey123
        maxLength: 50
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for the SOS Kotopes project. It provides endpoints
    for managing the database of animals in need of help.
  title: SOS Kotopes API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login through username and password
      operationId: login-basic
      parameters:
      - description: User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Login through username and password
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Signup through username and password
      operationId: signup
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: Firstname
        in: formData
        name: firstname
        type: string
      - description: Lastname
        in: formData
        name: lastname
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Photo
        in: formData
        name: photo
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Signup through username and password
      tags:
      - auth
  /auth/token/refresh:
    post:
      consumes:
      - application/json
      description: Refresh token
      operationId: refresh-token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuthCookie: []
      summary: Refresh token
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuthBasic:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"